+++ providers/org/zenodo/datacite/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'providers.org.zenodo.datacite.apps.AppConfig'
