--- providers/org/zenodo/oai/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = 'providers.org.zenodo.oai.apps.AppConfig'
